Working contrast calculation (WCAG 2.1)
