+++ src/components/ColorMatrix.css
@@ -1,12 +1,22 @@
-#color-matrix {
-  grid-column: 2;
+table {
   grid-row: 2;
-  display: grid;
-  grid-template-columns: repeat(auto-fit);
+  grid-column: 2;
 }
 
-td, th {
-  width: auto;
-  height: auto;
-  min-width: 24px;
+thead, tbody {
+  display: grid;
+  grid-auto-flow: row;
+  grid-auto-rows: 1fr;
+}
+
+tr {
+  display: grid;
+  grid-auto-flow: column;
+  grid-auto-columns: 1fr;
+  align-items: center;
+}
+
+th,
+td {
+  padding: 4px;
 }
